+++ generators/app/index.js
@@ -67,6 +67,7 @@
 
 		configuring() {
 			logs.priority(this, 'configuring', false);
+			this.currentVersion = '0.0.0';
 		}
 
 		inputAppName() {