+++ generators/controller/writing/controllerFile.js
@@ -21,7 +21,8 @@
 			controllerName: $that.controllerName + 'Controller',
 			nowDate       : $that.nowDate,
 			nowTime       : $that.nowTime,
-			authorShort   : $that.authorShort || $that.config.get('authorShort')
+			authorShort   : $that.authorShort || $that.config.get('authorShort'),
+			currentVersion: $that.currentVersion
 		};
 
 		utils.copyFileTplFrom($that, 'controller.txt', config, filePrefix + 'controller.js');