+++ generators/service/writing/serviceFile.js
@@ -14,14 +14,15 @@
 
 	module.exports = $that => {
 		const defaultPath = 'scripts/services';
-		const filePrefix = 'app/' + ($that.servicePath || defaultPath) + '/' + $that.serviceName + '.';
-		const config     = {
-			appNameCamel: $that.config.get('appNameCamel'),
-			appNameKebab: $that.config.get('appNameKebab'),
-			serviceName : $that.serviceName + 'Service',
-			nowDate     : $that.nowDate,
-			nowTime     : $that.nowTime,
-			authorShort : $that.authorShort || $that.config.get('authorShort')
+		const filePrefix  = 'app/' + ($that.servicePath || defaultPath) + '/' + $that.serviceName + '.';
+		const config      = {
+			appNameCamel  : $that.config.get('appNameCamel'),
+			appNameKebab  : $that.config.get('appNameKebab'),
+			serviceName   : $that.serviceName + 'Service',
+			nowDate       : $that.nowDate,
+			nowTime       : $that.nowTime,
+			authorShort   : $that.authorShort || $that.config.get('authorShort'),
+			currentVersion: $that.currentVersion
 		};
 
 		utils.copyFileTplFrom($that, 'service.txt', config, filePrefix + 'service.js');