+++ generators/view/writing/languagesFiles.js
@@ -22,7 +22,8 @@
 			appNameKebab       : $that.config.get('appNameKebab'),
 			nowDate            : $that.nowDate,
 			nowTime            : $that.nowTime,
-			authorShort        : $that.authorShort || $that.config.get('authorShort')
+			authorShort        : $that.authorShort || $that.config.get('authorShort'),
+			currentVersion     : $that.currentVersion
 		};
 
 		utils.copyFileTpl($that, 'languages/fr.json', config, filePrefix + 'fr.json');